requirements.txt
alpaca-trade-api==3.0.0
pandas==2.2.0
numpy==1.26.0
pyyaml==6.0
requests==2.31.0
schedule==1.2.1
openai==1.14.0

